requirements.txt
python-telegram-bot==10.0.1
gunicorn==19.7.1
requests==2.18.4

Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.0.3
Flask-Script==2.0.5
MySQL-python==1.2.5
